Add JUnit Tests
