+++ build.gradle
@@ -8,8 +8,6 @@ repositories {
     mavenCentral()
 }
 
-sourceSets.main.java.srcDirs = ['src']
-
 dependencies {
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.10.0'
     testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.10.0'
@@ -30,7 +28,7 @@ test {
 }
 
 application {
-    mainClass = "src.alpha.Alpha"
+    mainClass = "alpha.Alpha"
 }
 
 shadowJar {
