+++ src/main/java/alpha/Deadline.java
@@ -1,4 +1,4 @@
-package src.alpha;
+package alpha;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
