+++ src/main/java/alpha/ExitCommand.java
@@ -1,4 +1,4 @@
-package src.alpha;
+package alpha;
 public class ExitCommand extends Command{
 
     public ExitCommand(TaskList taskList, FileHandler fh, UI ui) {
