+++ src/main/java/alpha/FileHandler.java
@@ -1,4 +1,4 @@
-package src.alpha;
+package alpha;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
