+++ src/main/java/alpha/InvalidFormatException.java
@@ -1,4 +1,4 @@
-package src.alpha;
+package alpha;
 // Exception thrown when a task is input with an invalid format
 public class InvalidFormatException extends TaskException{
