+++ src/main/java/alpha/InvalidInputException.java
@@ -1,4 +1,4 @@
-package src.alpha;
+package alpha;
 // Exception thrown when an input is not recognized
 public class InvalidInputException extends AlphaException{
     public InvalidInputException(String errorMessage) {
