+++ src/main/java/alpha/MarkCommand.java
@@ -1,4 +1,4 @@
-package src.alpha;
+package alpha;
 public class MarkCommand extends Command{
 
     private int index;
