+++ src/main/java/alpha/MissingIndexException.java
@@ -1,4 +1,4 @@
-package src.alpha;
+package alpha;
 // Exception thrown when using a missing index to mark, delete or unmark a task
 public class MissingIndexException extends AlphaException{
     public MissingIndexException(String errorMessage) {
