+++ src/main/java/alpha/MissingInfoException.java
@@ -1,4 +1,4 @@
-package src.alpha;
+package alpha;
 // Exception that is thrown when a task is inputted with missing information
 public class MissingInfoException extends TaskException{
