+++ src/main/java/alpha/TaskException.java
@@ -1,4 +1,4 @@
-package src.alpha;
+package alpha;
 // Class to handle exceptions with entering tasks
 public abstract class TaskException extends AlphaException{
     public enum TaskType {
