+++ src/main/java/alpha/TaskList.java
@@ -1,4 +1,4 @@
-package src.alpha;
+package alpha;
 import java.util.ArrayList;
